requirements.txt
pandas>=2.0.0
yfinance>=0.2.36
forex-python>=1.8
requests>=2.31.0
beautifulsoup4>=4.12.0
tabulate>=0.9.0
openpyxl>=3.1.2
python-dateutil>=2.8.2
lxml>=4.9.3
numbers-parser>=4.15.1
PyYAML>=6.0
pyxirr>=0.10.0 